% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hook.R
\name{create_hook}
\alias{create_hook}
\title{Create a pre-commit hook}
\usage{
create_hook(id, name, description, language = "system", always_run = TRUE)
}
\arguments{
\item{id}{The unique identifier for the hook.}

\item{name}{A descriptive name for the hook.}

\item{description}{A brief description of what the hook does.}

\item{language}{The programming language or environment for the hook (default is "system").}

\item{always_run}{Logical, whether the hook should always run (default is TRUE).}
}
\description{
Add a pre-commit hook to .pre-commit-config.y*ml
}
