% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_precommit.R
\name{create_precommit_config}
\alias{create_precommit_config}
\alias{edit_precommit_config}
\title{Handle .pre-commit-config file}
\usage{
create_precommit_config(filename = path_precommit_files()[1], force = FALSE)

edit_precommit_config()
}
\arguments{
\item{filename}{the name of the file to create}

\item{force}{overwrite the file if it already exists}
}
\value{
cli messages related to the creation and edition of the \code{.pre-commit-config} file.
}
\description{
Handle .pre-commit-config file for running pre-commit hooks
}
\details{
\code{create_precommit_config()} will create a .pre-commit-config file in the current project.
Only one file is allowed in the project and should be in the root directory or in the inst directory such as:
\itemize{
\item inst/pre-commit/.pre-commit-config.yml
\item inst/pre-commit/.pre-commit-config.yaml
}
}
