% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_precommit_file.R
\name{create_precommit_file}
\alias{create_precommit_file}
\alias{edit_precommit_file}
\title{Handle .pre-commit-config file}
\usage{
create_precommit_file(filename = path_precommit_files(), force = FALSE)

edit_precommit_file()
}
\arguments{
\item{filename}{the name of the file to create}

\item{force}{overwrite the file if it already exists}
}
\description{
Handle .pre-commit-config file for running pre-commit hooks
}
\details{
\code{create_precommit_file()} will create a .pre-commit-config file in the current project.
Only one file is allowed in the project and should be in the root directory or in the inst directory such as:
\itemize{
\item .pre-commit-config.yml
\item .pre-commit-config.yaml
\item inst/.pre-commit-config.yml
\item inst/.pre-commit-config.yaml
}
}
