% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_precommit.R
\name{install_precommit}
\alias{install_precommit}
\title{Install pre-commit}
\usage{
install_precommit(force = FALSE)
}
\arguments{
\item{force}{overwrite the file if it already exists}
}
\description{
This function installs pre-commit file for the project.
}
