% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_precommit.R
\name{install_precommit}
\alias{install_precommit}
\title{Install pre-commit}
\usage{
install_precommit(force = FALSE, ...)
}
\arguments{
\item{force}{overwrite the file if it already exists}

\item{...}{additional arguments to pass to \code{create_precommit_config()}}
}
\value{
cli messages related to the installation of pre-commit files. Create \code{inst/pre-commit} folder and \code{.git/hooks/pre-commit} file.
}
\description{
Install pre-commit file in the repo.
}
