% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_precommit.R
\name{run_precommit}
\alias{run_precommit}
\title{Run precommit}
\usage{
run_precommit(path = ".git/hooks/pre-commit")
}
\arguments{
\item{path}{path of the pre-commit file}
}
\value{
cli message related to the run of all git precommit hooks.
}
\description{
Run pre-commit hooks
}
